Rename util.ts to utils.ts in OData handler
