+++ packages/handlers/odata/src/utils.ts
@@ -1,5 +1,3 @@
-import { ResolverData } from "@graphql-mesh/types";
-
 export function getUrlString(url: URL) {
   return decodeURIComponent(url.toString()).split('+').join(' ');
 }
